Partial bootstrap CA

An S4 class to store partial bootstrap correspondance analysis results.

Slots

id

A character string specifying the unique identifier of the corresponding matrix (UUID v4).

rows

A list of length three giving the vertices coordinates (x, y) of the samples convex hull and a identifier (id) to link each row to a sample.

columns

A list of length three giving the vertices coordinates (x, y) of the variables convex hull and a identifier (id) to link each row to a variable.

lengths

A list of two named numeric vectors giving the convex hull maximum dimension length of samples and variables, respectively.

cutoff

A length-two numeric vector giving the cutoff value for samples and variables selection, respectively.

keep

A list of two integer vectors giving the subscript of the samples and variables to be kept, respectively.

Subset

In the code snippets below, x is a BootCA object.

x[i, j, drop]

Extracts informations from a slot selected by subscript i thru j (see examples). i should be one of "rows" or "columns". Any unambiguous substring can be given. j is a numeric, integer or character vector or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the name of the elements. An empty index (a comma separated blank) indicates that all entries in that dimension are selected.

x[[i]]

Extracts informations from a slot selected by subscript i. i should be one of "id", "rows", "columns", "lengths", "cutoff" or "keep". Any unambiguous substring can be given.