Date model

An S4 class to store the event and accumulation times of archaeological assemblages as well as the results of resampling methods for date model checking.

# S4 method for DateModel
[[(x, i)

# S4 method for DateModel
[(x, i, j, drop = TRUE)

Arguments

x
A DateModel object from which to extract element(s).

i, j
Indices specifying elements to extract. i is a character string matching to the name of a slot. j can be missing or NULL, a numeric or character vector. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object.

drop
A logical scalar: should the result be coerced to the lowest possible dimension?

Methods (by generic)

[[: Extracts informations from a slot selected by subscript i.

[: Allows to select a slot thru j.

Slots

counts

A numeric matrix of count data.

dates

A two columns data.frame giving the known dates used for model fitting and an identifier to link each row to an assemblage.

level

A length-one numeric vector giving the confidence level.

model

A multiple linear model: the Gaussian multiple linear regression model fitted for event date estimation and prediction.

residual

A length-one numeric vector giving the residual standard deviation.

rows

A five columns data.frame giving the predicted event dates for each archaeological assemblage, with the following columns:

id

An identifier to link each row to an assemblage.

estimation

The event date estimation.

earliest

The lower boundary of the confidence interval.

latest

The upper boundary of the confidence interval.

error

The standard error of predicted dates.

columns

A five columns data.frame giving the predicted event dates for each archaeological type or fabric, with the following columns:

id

An identifier to link each row to an assemblage.

estimation

The event date estimation.

earliest

The lower boundary of the confidence interval.

latest

The upper boundary of the confidence interval.

error

The standard error of predicted dates.

accumulation

A two columns data.frame giving the point estimate of accumulation dates of archaeological assemblages and an identifier to link each row to an assemblage.

jackknife

A six columns data.frame giving the results of the resamping procedure (jackknifing fabrics) for each assemblage (in rows) with the following columns:

id

An identifier to link each row to an assemblage.

estimation

The jackknife event date estimate.

earliest

The lower boundary of the associated prediction interval.

latest

The upper boundary of the associated prediction interval.

error

The standard error of predicted means.

bias

The jackknife estimate of bias.

bootstrap

A six columns data.frame giving the boostrap distribution statistics for each replicated assemblage (in rows) with the following columns:

id

An identifier to link each row to an assemblage.

min

Minimum value.

Q05

Sample quantile to 0.05 probability.

mean

Mean value (event date).

Q95

Sample quantile to 0.95 probability.

max

Maximum value.