Bootstrap Estimation

Samples randomly from the elements of object with replacement.

Usage

# S4 method for DiversityIndex
bootstrap(object, n = 1000, f = NULL)

Arguments

object

An R object (typically a DiversityIndex object).

n

A non-negative integer giving the number of bootstrap replications.

f

A function that takes a single numeric vector (the result of do) as argument.

Value

If f is NULL (the default), bootstrap() returns a named numeric

vector with the following elements:

original

The observed value of do applied to object.

mean

The bootstrap estimate of mean of do.

bias

The bootstrap estimate of bias of do.

error

he bootstrap estimate of standard error of do.

If f is a function, bootstrap() returns the result of f applied to the n values of do.

See also

Author

N. Frerebeau

Examples

## Data from Conkey 1980, Kintigh 1989
data("cantabria")

## Shannon diversity index
(h <- heterogeneity(cantabria, method = "shannon"))
#> [1] 3.269200 2.955298 2.491683 2.485604 2.329187

## Bootstrap resampling
bootstrap(h, f = NULL)
#>                  original     mean         bias      error
#> Altamira         3.269200 3.277545  0.008345660 0.07711876
#> Cueto de la Mina 2.955298 2.971006  0.015708141 0.14682904
#> El Juyo          2.491683 2.521278  0.029594905 0.19339010
#> El Cierro        2.485604 2.475748 -0.009855441 0.22689826
#> La Paloma        2.329187 2.310532 -0.018654700 0.26595783

bootstrap(h, f = summary)
#>                       Min.  1st Qu.   Median     Mean  3rd Qu.     Max.
#> Altamira         2.9766019 3.224816 3.282332 3.277187 3.332554 3.522391
#> Cueto de la Mina 2.3885562 2.883319 2.984141 2.976524 3.082799 3.323453
#> El Juyo          1.6726254 2.395326 2.529993 2.518629 2.655832 3.210391
#> El Cierro        1.4391741 2.361293 2.501051 2.492517 2.651700 3.024747
#> La Paloma        0.5623351 2.150060 2.334882 2.306934 2.499212 2.969014

quant <- function(x) quantile(x, probs = c(0.25, 0.50))
bootstrap(h, f = quant)
#>                       25%      50%
#> Altamira         3.228015 3.283690
#> Cueto de la Mina 2.881741 2.983222
#> El Juyo          2.400641 2.533444
#> El Cierro        2.353114 2.497983
#> La Paloma        2.153532 2.350159

## Jackknife resampling
jackknife(h)
#>                  original     mean       bias      error
#> Altamira         3.269200 3.246457 -0.9779158 0.08145529
#> Cueto de la Mina 2.955298 2.932856 -0.9649994 0.15907033
#> El Juyo          2.491683 2.469588 -0.9500688 0.19670002
#> El Cierro        2.485604 2.462723 -0.9838709 0.22195573
#> La Paloma        2.329187 2.305876 -1.0023704 0.25266291

bootstrap(h, f = summary)
#>                      Min.  1st Qu.   Median     Mean  3rd Qu.     Max.
#> Altamira         2.990321 3.226704 3.284850 3.281248 3.339730 3.488504
#> Cueto de la Mina 2.360903 2.876108 2.974649 2.973087 3.084857 3.362741
#> El Juyo          1.761097 2.398572 2.529897 2.516077 2.646910 3.042476
#> El Cierro        1.468140 2.334882 2.503830 2.484055 2.641085 3.048770
#> La Paloma        1.153742 2.162710 2.345737 2.323236 2.507696 2.886470