Heterogeneity and Evenness

index_heterogeneity returns an heterogeneity or dominance index.

index_evenness returns an evenness measure.

index_heterogeneity(object, ...)

index_evenness(object, ...)

# S4 method for CountMatrix
index_evenness(
  object,
  method = c("shannon", "brillouin", "mcintosh", "simpson"),
  simulate = FALSE,
  quantiles = TRUE,
  level = 0.8,
  step = 1,
  n = 1000,
  ...
)

# S4 method for CountMatrix
index_heterogeneity(
  object,
  method = c("berger", "brillouin", "mcintosh", "shannon", "simpson"),
  simulate = FALSE,
  quantiles = TRUE,
  level = 0.8,
  step = 1,
  n = 1000,
  ...
)

Arguments

object
A \(m \times p\) matrix of count data (typically a CountMatrix object).

...
Further arguments to be passed to internal methods.

method
A character string specifying the index to be computed (see details). Any unambiguous substring can be given.

simulate
A logical scalar: simulated assemblages be computed?

quantiles
A logical scalar: should sample quantiles be used as confidence interval? If TRUE (the default), sample quantiles are used as described in Kintigh (1989), else quantiles of the normal distribution are used. Only used if simulate is TRUE.

level
A length-one numeric vector giving the confidence level. Only used if simulate is TRUE.

step
A non-negative integer giving the increment of the sample size. Only used if simulate is TRUE.

n
A non-negative integer giving the number of bootstrap replications. Only used if simulate is TRUE.

Value

index_heterogeneity and index_evenness return a DiversityIndex object.

Details

Diversity measurement assumes that all individuals in a specific taxa are equivalent and that all types are equally different from each other (Peet 1974). A measure of diversity can be achieved by using indices built on the relative abundance of taxa. These indices (sometimes referred to as non-parametric indices) benefit from not making assumptions about the underlying distribution of taxa abundance: they only take relative abundances of the species that are present and species richness into account. Peet (1974) refers to them as indices of heterogeneity.

Diversity indices focus on one aspect of the taxa abundance and emphasize either richness (weighting towards uncommon taxa) or dominance (weighting towards abundant taxa; Magurran 1988).

Evenness is a measure of how evenly individuals are distributed across the sample.

The following heterogeneity index and corresponding evenness measures are available (see Magurran 1988 for details):

berger

Berger-Parker dominance index. The Berger-Parker index expresses the proportional importance of the most abundant type. This metric is highly biased by sample size and richness, moreover it does not make use of all the information available from sample.

brillouin

Brillouin diversity index. The Brillouin index describes a known collection: it does not assume random sampling in an infinite population. Pielou (1975) and Laxton (1978) argues for the use of the Brillouin index in all circumstances, especially in preference to the Shannon index.

mcintosh

McIntosh dominance index. The McIntosh index expresses the heterogeneity of a sample in geometric terms. It describes the sample as a point of a S-dimensional hypervolume and uses the Euclidean distance of this point from the origin.

shannon

Shannon-Wiener diversity index. The Shannon index assumes that individuals are randomly sampled from an infinite population and that all taxa are represented in the sample (it does not reflect the sample size). The main source of error arises from the failure to include all taxa in the sample: this error increases as the proportion of species discovered in the sample declines (Peet 1974, Magurran 1988). The maximum likelihood estimator (MLE) is used for the relative abundance, this is known to be negatively biased by sample size.

simpson

Simpson dominance index for finite sample. The Simpson index expresses the probability that two individuals randomly picked from a finite sample belong to two different types. It can be interpreted as the weighted mean of the proportional abundances. This metric is a true probability value, it ranges from 0 (perfectly uneven) to 1 (perfectly even).

The berger, mcintosh and simpson methods return a dominance index, not the reciprocal or inverse form usually adopted, so that an increase in the value of the index accompanies a decrease in diversity.

Note

Ramanujan approximation is used for \(x!\) computation if \(x > 170\).

References

Berger, W. H. & Parker, F. L. (1970). Diversity of Planktonic Foraminifera in Deep-Sea Sediments. Science, 168(3937), 1345-1347. DOI: 10.1126/science.168.3937.1345.

Brillouin, L. (1956). Science and information theory. New York: Academic Press.

Kintigh, K. W. (1984). Measuring Archaeological Diversity by Comparison with Simulated Assemblages. American Antiquity, 49(1), 44-54. DOI: 10.2307/280511.

Kintigh, K. W. (1989). Sample Size, Significance, and Measures of Diversity. In Leonard, R. D. and Jones, G. T., Quantifying Diversity in Archaeology. New Directions in Archaeology. Cambridge: Cambridge University Press, p. 25-36.

Laxton, R. R. (1978). The measure of diversity. Journal of Theoretical Biology, 70(1), 51-67. DOI: 10.1016/0022-5193(78)90302-8.

Magurran, A. E. (1988). Ecological Diversity and its Measurement. Princeton, NJ: Princeton University Press. DOI: 10.1007/978-94-015-7358-0.

McIntosh, R. P. (1967). An Index of Diversity and the Relation of Certain Concepts to Diversity. Ecology, 48(3), 392-404. DOI: 10.2307/1932674.

Peet, R. K. (1974). The Measurement of Species Diversity. Annual Review of Ecology and Systematics, 5(1), 285-307. DOI: 10.1146/annurev.es.05.110174.001441.

Pielou, E. C. (1975). Ecological Diversity. New York: Wiley. DOI: 10.4319/lo.1977.22.1.0174b

Shannon, C. E. (1948). A Mathematical Theory of Communication. The Bell System Technical Journal, 27, 379-423. DOI: 10.1002/j.1538-7305.1948.tb01338.x.

Simpson, E. H. (1949). Measurement of Diversity. Nature, 163(4148), 688-688. DOI: 10.1038/163688a0.

See also

Author

N. Frerebeau

Examples

## Shannon diversity index
## Data from Magurran 1988, p. 145-149
birds <- CountMatrix(
  data = c(35, 26, 25, 21, 16, 11, 6, 5, 3, 3,
           3, 3, 3, 2, 2, 2, 1, 1, 1, 1, 0, 0,
           30, 30, 3, 65, 20, 11, 0, 4, 2, 14,
           0, 3, 9, 0, 0, 5, 0, 0, 0, 0, 1, 1),
  nrow = 2, byrow = TRUE, dimnames = list(c("oakwood", "spruce"), NULL))

index_heterogeneity(birds, "shannon") # 2.40 2.06

#> <HeterogeneityIndex: shannon>
#>         size    index
#> oakwood  170 2.407983
#> spruce   198 2.055991

index_evenness(birds, "shannon") # 0.80 0.78

#> <EvennessIndex: shannon>
#>         size     index
#> oakwood  170 0.8038044
#> spruce   198 0.7790626

## Brillouin diversity index
## Data from Magurran 1988, p. 150-151
moths <- CountMatrix(data = c(17, 15, 11, 4, 4, 3, 3, 3, 2, 2, 1, 1, 1),
                     nrow = 1, byrow = TRUE)

index_heterogeneity(moths, "brillouin") # 1.88

#> <HeterogeneityIndex: brillouin>
#>      size    index
#> row1   67 1.876307

index_evenness(moths, "brillouin") # 0.83

#> <EvennessIndex: brillouin>
#>      size     index
#> row1   67 0.8275138

## Simpson dominance index
## Data from Magurran 1988, p. 152-153
trees <- CountMatrix(
  data = c(752, 276, 194, 126, 121, 97, 95, 83, 72, 44, 39,
           16, 15, 13, 9, 9, 9, 8, 7, 4, 2, 2, 1, 1, 1),
  nrow = 1, byrow = TRUE
)

index_heterogeneity(trees, "simpson") # 1.19

#> <HeterogeneityIndex: simpson>
#>      size    index
#> row1 1996 0.186547

index_evenness(trees, "simpson") # 0.21

#> <EvennessIndex: simpson>
#>      size     index
#> row1 1996 0.2144231

## McIntosh dominance index
## Data from Magurran 1988, p. 154-155
invertebrates <- CountMatrix(
  data = c(254, 153, 90, 69, 68, 58, 51, 45, 40, 39, 25, 23, 19, 18, 16, 14, 14,
           11, 11, 11, 11, 10, 6, 6, 6, 6, 5, 3, 3, 3, 3, 3, 1, 1, 1, 1, 1, 1),
  nrow = 1, byrow = TRUE
)

index_heterogeneity(invertebrates, "mcintosh") # 0.71

#> <HeterogeneityIndex: mcintosh>
#>      size     index
#> row1 1100 0.7066343

index_evenness(invertebrates, "mcintosh") # 0.82

#> <EvennessIndex: mcintosh>
#>      size     index
#> row1 1100 0.8180305

## Berger-Parker dominance index
## Data from Magurran 1988, p. 156-157
fishes <- CountMatrix(
  data = c(394, 3487, 275, 683, 22, 1, 0, 1, 6, 8, 1, 1, 2,
           1642, 5681, 196, 1348, 12, 0, 1, 48, 21, 1, 5, 7, 3,
           90, 320, 180, 46, 2, 0, 0, 1, 0, 0, 2, 1, 5,
           126, 17, 115, 436, 27, 0, 0, 3, 1, 0, 0, 1, 0,
           32, 0, 0, 5, 0, 0, 0, 0, 13, 9, 0, 0, 4),
  nrow = 5, byrow = TRUE,
  dimnames = list(c("station 1", "station 2", "station 3",
                    "station 4", "station 5"), NULL)
)

index_heterogeneity(fishes, "berger") # 0.71 0.63 0.50 0.60 0.51

#> <HeterogeneityIndex: berger>
#>           size     index
#> station 1 4881 0.7144028
#> station 2 8965 0.6336866
#> station 3  647 0.4945904
#> station 4  726 0.6005510
#> station 5   63 0.5079365