Diversity, dominance and evenness index

dominanceBerger returns Berger-Parker dominance index. diversityBrillouin and evennessBrillouin return Brillouin diversity index and evenness. dominanceMcintosh and evennessMcintosh return McIntosh dominance index and evenness. diversityShannon, evennessShannon, varianceShannon return Shannon-Wiener diversity index, evenness and variance. dominanceSimpson and evennessSimpson return Simpson dominance index and evenness.

dominanceBerger(x, ...)

diversityBrillouin(x, ...)

evennessBrillouin(x, ...)

dominanceMcintosh(x, ...)

evennessMcintosh(x, ...)

diversityShannon(x, base = exp(1), ...)

evennessShannon(x, ...)

varianceShannon(x, ...)

dominanceSimpson(x, ...)

evennessSimpson(x, ...)

Arguments

x
A numeric vector giving the number of individuals for each class.

...
Currently not used.

Value

A length-one numeric vector.

References

Berger, W. H. & Parker, F. L. (1970). Diversity of Planktonic Foraminifera in Deep-Sea Sediments. Science, 168(3937), 1345-1347. DOI: 10.1126/science.168.3937.1345.

Brillouin, L. (1956). Science and information theory. New York: Academic Press.

McIntosh, R. P. (1967). An Index of Diversity and the Relation of Certain Concepts to Diversity. Ecology, 48(3), 392-404. DOI: 10.2307/1932674.

Shannon, C. E. (1948). A Mathematical Theory of Communication. The Bell System Technical Journal, 27, 379-423. DOI: 10.1002/j.1538-7305.1948.tb01338.x.

Simpson, E. H. (1949). Measurement of Diversity. Nature, 163(4148), 688-688. DOI: 10.1038/163688a0.