Richness index

Abundance data: richnessACE returns Abundance-based Coverage Estimator (ACE). richnessChao1 returns (improved) Chao1 estimator for abundance data. richnessMargalef returns Margalef richness index. richnessMenhinick returns Menhinick richness index.

richnessACE(x, k = 10, ...)

richnessChao1(x, unbiased = FALSE, improved = FALSE, ...)

richnessMargalef(x, ...)

richnessMenhinick(x, ...)

richnessICE(x, k = 10, ...)

richnessChao2(x, unbiased = FALSE, improved = FALSE, ...)

Arguments

x
A numeric vector giving the number of individuals for each type (abundance data) or a logical matrix (replicated incidence data).

k
A numeric vector giving the threshold between rare and abundant species.

...
Currently not used.

unbiased
A logical scalar. Should the bias-corrected estimator be used?

improved
A logical scalar. Should the improved estimator be used?

Value

A length-one numeric vector.

Details

Incidence data: richnessICE returns Incidence-based Coverage Estimator (ICE). richnessChao2 returns (improved) Chao2 estimator for replicated incidence data.

References

Chao, A. (1984). Nonparametric Estimation of the Number of Classes in a Population. Scandinavian Journal of Statistics, 11(4), 265-270.

Chao, A. (1987). Estimating the Population Size for Capture-Recapture Data with Unequal Catchability. Biometrics 43(4): 783. DOI: 10.2307/2531532.

Chao, A., & Lee, S.-M. (1992). Estimating the Number of Classes via Sample Coverage. Journal of the American Statistical Association, 87(417), 210-217. DOI: 10.1080/01621459.1992.10475194

Chiu, C.-H., Wang, Y.-T., Walther, B. A. & Chao, A. (2014). An improved nonparametric lower bound of species richness via a modified good-turing frequency formula. Biometrics, 70(3), 671-682. DOI: 10.1111/biom.12200.

Margalef, R. (1958). Information Theory in Ecology. General Systems, 3, 36-71.

Menhinick, E. F. (1964). A Comparison of Some Species-Individuals Diversity Indices Applied to Samples of Field Insects. Ecology, 45(4), 859-861. DOI: 10.2307/1934933.

See also