Rarefaction

Computes Hurlbert's unbiased estimate of Sander's rarefaction.

Usage

rarefaction(object, ...)

index_hurlbert(x, ...)

# S4 method for matrix
rarefaction(object, sample = NULL, method = c("hurlbert"), step = 1)

# S4 method for data.frame
rarefaction(object, sample = NULL, method = c("hurlbert"), step = 1)

# S4 method for numeric
index_hurlbert(x, sample, ...)

Arguments

object

A \(m \times p\) numeric matrix or data.frame of count data (absolute frequencies).

...

Currently not used.

x

A numeric vector of count data (absolute frequencies).

sample

A length-one numeric vector giving the sub-sample size. The size of sample should be smaller than total community size.

method

A character string or vector of strings specifying the index to be computed (see details). Any unambiguous substring can be given.

step

An integer giving the increment of the sample size.

Value

rarefaction() returns a RarefactionIndex object.

index_*() return a numeric vector.

Details

The number of different taxa, provides an instantly comprehensible expression of diversity. While the number of taxa within a sample is easy to ascertain, as a term, it makes little sense: some taxa may not have been seen, or there may not be a fixed number of taxa (e.g. in an open system; Peet 1974). As an alternative, richness (\(S\)) can be used for the concept of taxa number (McIntosh 1967).

It is not always possible to ensure that all sample sizes are equal and the number of different taxa increases with sample size and sampling effort (Magurran 1988). Then, rarefaction (\(E(S)\)) is the number of taxa expected if all samples were of a standard size (i.e. taxa per fixed number of individuals). Rarefaction assumes that imbalances between taxa are due to sampling and not to differences in actual abundances.

References

Hurlbert, S. H. (1971). The Nonconcept of Species Diversity: A Critique and Alternative Parameters. Ecology, 52(4), 577-586. doi:10.2307/1934145 .

Sander, H. L. (1968). Marine Benthic Diversity: A Comparative Study. The American Naturalist, 102(925), 243-282.

See also

Author

N. Frerebeau

Examples

## Richness
## Margalef and Menhinick index
## Data from Magurran 1988, p. 128-129
trap <- matrix(data = c(9, 3, 0, 4, 2, 1, 1, 0, 1, 0, 1, 1,
                        1, 0, 1, 0, 0, 0, 1, 2, 0, 5, 3, 0),
               nrow = 2, byrow = TRUE, dimnames = list(c("A", "B"), NULL))
richness(trap, method = "margalef") # 2.55 1.88
#> [1] 2.551432 1.949356
richness(trap, method = "menhinick") # 1.95 1.66
#> [1] 1.876630 1.664101

## Asymptotic species richness
## Chao1-type estimators
## Data from Chao & Chiu 2016
brazil <- matrix(
  data = rep(x = c(1:21, 23, 25, 27, 28, 30, 32, 34:37, 41,
                   45, 46, 49, 52, 89, 110, 123, 140),
             times = c(113, 50, 39, 29, 15, 11, 13, 5, 6, 6, 3, 4,
                       3, 5, 2, 5, 2, 2, 2, 2, 1, 2, 1, 1, 1, 1, 1,
                       0, 0, 2, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0)),
  nrow = 1, byrow = TRUE
)

composition(brazil, method = c("chao1"), unbiased = FALSE) # 461.625
#> [1] 461.6254
composition(brazil, method = c("ace"), k = 10) # 445.822
#> [1] 445.8224

## Rarefaction
rarefaction(trap, sample = 13) # 6.56 6.00
#>   1        2      3        4        5        6        7        8        9
#> A 1 1.818182 2.5048 3.095765 3.616393 4.084312 4.511693 4.906919 5.275811
#> B 1 1.820513 2.5000 3.069930 3.555556 3.976690 4.348485 4.682984 4.986014
#>         10       11       12       13
#> A 5.622502 5.950041 6.260802 6.556735
#> B 5.269231 5.538462 6.000000 6.000000