Resample Diversity

Performs bootstrap/jackknife resampling.

Usage

# S4 method for DiversityIndex
bootstrap(
  object,
  level = 0.95,
  type = c("student", "normal"),
  probs = c(0.25, 0.5, 0.75),
  n = 1000
)

# S4 method for DiversityIndex
jackknife(object)

Arguments

object

A numeric or an integer vector or a CA or PCA object (see below).

level

A length-one numeric vector giving the confidence level. Must be a single number between \(0\) and \(1\). If NULL, no confidence interval are computed.

type

A character string giving the type of confidence interval to be returned. It must be one "student" (default) or "normal". Any unambiguous substring can be given. Only used if level is not NULL.``

probs

A numeric vector of probabilities with values in \([0,1]\) (see stats::quantile()). If NULL, quantiles are not computed.

n

A non-negative integer giving the number of bootstrap replications.

Value

Returns a data.frame.

See also

Author

N. Frerebeau

Examples

## Coerce dataset to a count matrix
data("chevelon", package = "folio")
chevelon <- as_count(chevelon)

## Shannon diversity index
(index_h <- index_heterogeneity(chevelon, method = "shannon"))
#> <HeterogeneityIndex: shannon>
#>       size     index
#> P610s   14 1.6681740
#> P610e    8 1.2130076
#> P625     4 0.6931472
#> P630     1 0.0000000
#> P307     1 0.0000000
#> P631    12 1.7481555
#> P623     3 1.0986123
#> P624    14 1.8711604
#> P626s   26 1.7207095
#> P626e  106 1.9115521
#> P627     4 1.0397208
#> P628    12 1.8200760
(index_e <- index_evenness(chevelon, method = "shannon"))
#> <EvennessIndex: shannon>
#>       size     index
#> P610s   14 0.8572718
#> P610e    8 0.8750000
#> P625     4 1.0000000
#> P630     1       NaN
#> P307     1       NaN
#> P631    12 0.8983742
#> P623     3 1.0000000
#> P624    14 0.9615862
#> P626s   26 0.8842698
#> P626e  106 0.8699849
#> P627     4 0.9463946
#> P628    12 0.9353340

## Bootstrap resampling
bootstrap(index_e)
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>             min      mean       max     lower     upper       Q25       Q50
#> P610s 0.5446150 0.8703724 0.9952326 0.8665131 0.8742318 0.8312828 0.8757609
#> P610e 0.5435644 0.8658888 1.0000000        NA        NA 0.8194484 0.8868595
#> P625  0.8112781 0.8941787 1.0000000        NA        NA 0.8112781 0.8112781
#> P630        Inf       NaN      -Inf        NA        NA        NA        NA
#> P307        Inf       NaN      -Inf        NA        NA        NA        NA
#> P631  0.5152734 0.8894606 1.0000000 0.8856168 0.8933044 0.8592975 0.8979445
#> P623  0.9182958 0.9372296 1.0000000        NA        NA 0.9182958 0.9182958
#> P624  0.7544450 0.9204857 0.9931223 0.9180914 0.9228799 0.8982265 0.9285249
#> P626s 0.6772627 0.8866834 0.9934151 0.8838576 0.8895091 0.8580158 0.8905079
#> P626e 0.7468703 0.8703223 0.9557807 0.8683887 0.8722558 0.8506856 0.8716096
#> P627  0.8112781 0.9078217 1.0000000        NA        NA 0.8112781 0.9463946
#> P628  0.7091479 0.9138856 1.0000000 0.9109813 0.9167899 0.8848617 0.9155385
#>             Q75
#> P610s 0.9134443
#> P610e 0.9528195
#> P625  1.0000000
#> P630         NA
#> P307         NA
#> P631  0.9355245
#> P623  0.9182958
#> P624  0.9517472
#> P626s 0.9190369
#> P626e 0.8921603
#> P627  0.9463946
#> P628  0.9439593

## Jackknife resampling
jackknife(index_e)
#>            mean          bias        error
#> P610s 0.8530134 -3.832613e-02 3.467900e-02
#> P610e 0.8725916 -2.167581e-02 1.055074e-01
#> P625        NaN           NaN          NaN
#> P630        NaN           NaN          NaN
#> P307        NaN           NaN          NaN
#> P631  0.8956783 -2.426312e-02 3.790426e-02
#> P623  1.0000000  9.992007e-16 3.330669e-16
#> P624  0.9598834 -1.532543e-02 1.972936e-02
#> P626s 0.8786773 -5.033240e-02 4.931785e-02
#> P626e 0.8645734 -4.870273e-02 4.770957e-02
#> P627  0.9461354 -2.333001e-03 6.329892e-02
#> P628  0.9329141 -2.177978e-02 1.598781e-02