Richness and rarefaction

richness returns sample richness. rarefaction returns Hurlbert's unbiaised estimate of Sander's rarefaction.

richness(object, ...)

rarefaction(object, ...)

# S4 method for CountMatrix
rarefaction(object, sample)

# S4 method for CountMatrix
richness(object, method = c("ace", "chao1",
  "chao1i", "margalef", "menhinick", "none"), unbiased = FALSE, k = 10,
  simplify = FALSE)

# S4 method for IncidenceMatrix
richness(object, method = c("chao2",
  "chao2i", "ice"), unbiased = FALSE, k = 10, simplify = FALSE)

Arguments

object
A \(m \times p\) matrix of count data.

...
Further arguments passed to other methods.

sample
A length-one numeric vector giving the sub-sample size.

method
A character string or vector of strings specifiying the index to be computed (see details). Any unambiguous substring can be given.

unbiased
A logical scalar. Should the bias-corrected estimator be used? Only used with "chao1" or "chao2" (improved) estimator.

k
A length-one numeric vector giving the threshold between rare/infrequent and abundant/frequent species. Only used if method is "ace" or "ice".

simplify
A logical scalar: should the result be simplified to a matrix? The default value, FALSE, returns a list.

Value

rarefaction returns a numeric vector.

If simplify is FALSE, then richness method returns a list (default), else returns a matrix (for CountMatrix) or a a numeric vector (for IncidenceMatrix).

Details

The number of different taxa, provides an instantly comprehensible expression of diversity. While the number of taxa within a sample is easy to ascertain, as a term, it makes little sense: some taxa may not have been seen, or there may not be a fixed number of taxa (e.g. in an open system; Peet 1974). As an alternative, richness (\(S\)) can be used for the concept of taxa number (McIntosh 1967).

It is not always possible to ensure that all sample sizes are equal and the number of different taxa increases with sample size and sampling effort (Magurran 1988). Then, rarefaction (\(E(S)\)) is the number of taxa expected if all samples were of a standard size (i.e. taxa per fixed number of individuals). Rarefaction assumes that imbalances between taxa are due to sampling and not to differences in actual abundances.

The following richness measures are available for count data:

ace

Abundance-based Coverage Estimator.

chao1

Chao1 estimator.

chao1i

Improved Chao1 estimator.

margalef

Margalef richness index.

menhinick

Menhinick richness index.

none

Returns the number of observed taxa/types.

The following richness measures are available for replicated incidence data:

ice

Incidence-based Coverage Estimator.

chao2

Chao2 estimator.

chao2i

Improved Chao2 estimator.

References

Chao, A. (1984). Nonparametric Estimation of the Number of Classes in a Population. Scandinavian Journal of Statistics, 11(4), 265-270.

Chao, A. (1987). Estimating the Population Size for Capture-Recapture Data with Unequal Catchability. Biometrics 43(4), 783-791. DOI: 10.2307/2531532.

Chao, A. & Chiu, C.-H. (2016). Species Richness: Estimation and Comparison. In Balakrishnan, N., Colton, T., Everitt, B., Piegorsch, B., Ruggeri, F. & Teugels, J. L. (Eds.), Wiley StatsRef: Statistics Reference Online. Chichester, UK: John Wiley & Sons, Ltd., 1-26. DOI: 10.1002/9781118445112.stat03432.pub2

Chao, A. & Lee, S.-M. (1992). Estimating the Number of Classes via Sample Coverage. Journal of the American Statistical Association, 87(417), 210-217. DOI: 10.1080/01621459.1992.10475194.

Chiu, C.-H., Wang, Y.-T., Walther, B. A. & Chao, A. (2014). An improved nonparametric lower bound of species richness via a modified good-turing frequency formula. Biometrics, 70(3), 671-682. DOI: 10.1111/biom.12200.

Hurlbert, S. H. (1971). The Nonconcept of Species Diversity: A Critique and Alternative Parameters. Ecology, 52(4), 577-586. DOI: 10.2307/1934145.

Magurran, A. E. (1988). Ecological Diversity and its Measurement. Princeton, NJ: Princeton University Press. DOI:10.1007/978-94-015-7358-0.

Margalef, R. (1958). Information Theory in Ecology. General Systems, 3, 36-71.

Menhinick, E. F. (1964). A Comparison of Some Species-Individuals Diversity Indices Applied to Samples of Field Insects. Ecology, 45(4), 859-861. DOI: 10.2307/1934933.

McIntosh, R. P. (1967). An Index of Diversity and the Relation of Certain Concepts to Diversity. Ecology, 48(3), 392-404. DOI: 10.2307/1932674.

Sander, H. L. (1968). Marine Benthic Diversity: A Comparative Study. The American Naturalist, 102(925), 243-282.

See also

Examples

# Richness
## Margalef and Menhinick index
## Data from Magurran 1988, p. 128-129
trap <- CountMatrix(data = c(9, 3, 0, 4, 2, 1, 1, 0, 1, 0, 1, 1,
                             1, 0, 1, 0, 0, 0, 1, 2, 0, 5, 3, 0),
                    nrow = 2, byrow = TRUE, dimnames = list(c("A", "B"), NULL))
richness(trap, method = c("margalef", "menhinick"), simplify = TRUE)

#>   margalef menhinick
#> A 2.551432  1.876630
#> B 1.949356  1.664101

## A 2.55 1.88
## B 1.95 1.66

## Chao1-type estimators
## Data from Chao & Chiu 2016
brazil <- CountMatrix(
  data = rep(x = c(1:21, 23, 25, 27, 28, 30, 32, 34:37, 41,
                   45, 46, 49, 52, 89, 110, 123, 140),
             times = c(113, 50, 39, 29, 15, 11, 13, 5, 6, 6, 3, 4,
                       3, 5, 2, 5, 2, 2, 2, 2, 1, 2, 1, 1, 1, 1, 1,
                       0, 0, 2, 1, 1, 1, 1, 1, 0, 1, 1, 0, 0)),
  nrow = 1, byrow = TRUE
)

richness(brazil, method = c("chao1", "chao1i", "ace"),
         unbiased = FALSE, simplify = TRUE)

#>  chao1.1 chao1i.1    ace.1 
#> 461.6254 488.2843 445.8224 

## 461.625 488.284 445.822

# Rarefaction
rarefaction(trap, sample = 13) # 6.56

#>        A        B 
#> 6.556735       NA