Bootstrap Estimation

Bootstrap Estimation

Usage

stats_bootstrap(x, do, probs = c(0.05, 0.95), n = 1000, na.rm = FALSE, ...)

Arguments

x

A vector.

do

A function that takes x as an argument and returns a single numeric value.

probs

A numeric vector of probabilities with values in \([0,1]\) (see stats::quantile()).

n

A non-negative integer giving the number of bootstrap replications.

na.rm

A logical scalar: should missing values be removed from x before the quantiles are computed?

...

Extra arguments passed to do.

Value

A numeric vector with the following elements:

min

Minimum value.

mean

Mean value.

max

Maximum value.

Q*

Sample quantile to * probability.