tabula: Analysis, Seriation and Visualization of Archaeological Count Data

An easy way to examine archaeological count data. This package provides a convenient and reproducible toolkit for relative and absolute dating and analysis of (chronological) patterns. It includes functions for matrix seriation (reciprocal ranking, CA-based seriation), chronological modeling and dating of archaeological assemblages and/or objects. Beyond these, the package provides several tests and measures of diversity: heterogeneity and evenness (Brillouin, Shannon, Simpson, etc.), richness and rarefaction (Chao1, Chao2, ACE, ICE, etc.), turnover and similarity (Brainerd-Robinson, etc.). The package make it easy to visualize count data and statistical thresholds: rank vs. abundance plots, heatmaps, Ford (1962) and Bertin (1977) diagrams.

Details

Package:	tabula

Type:	Package

Version:	1.3.0

License:	GPL-3

DOI:	10.5281/zenodo.1489944

Note

This work received a state financial support managed by the Agence Nationale de la Recherche (France) throught the program Investissements d'avenir (ref. ANR-10-LABX-52).

See also