Turnover

Returns the degree of turnover in taxa composition along a grandient or transect.

turnover(object, ...)

# S4 method for CountMatrix
turnover(object, method = c("whittaker", "cody",
  "routledge1", "routledge2", "routledge3", "wilson"), simplify = FALSE,
  ...)

# S4 method for IncidenceMatrix
turnover(object, method = c("whittaker",
  "cody", "routledge1", "routledge2", "routledge3", "wilson"),
  simplify = FALSE, ...)

Arguments

object
A \(m \times p\) matrix of count data.

...
Further arguments passed to other methods.

method
A character string specifiying the method to be used (see details). Any unambiguous substring can be given.

simplify
A logical scalar: should the result be simplified to a matrix? The default value, FALSE, returns a list.

Value

If simplify is FALSE, returns a list (default), else returns a matrix.

Details

The following methods can be used to acertain the degree of turnover in taxa composition along a gradient (\(\beta\)-diversity) on qualitative (presence/absence) data. This assumes that the order of the matrix rows (from 1 to \(n\)) follows the progression along the gradient/transect.

whittaker

Whittaker measure.

cody

Cody measure.

routledge1

Routledge first measure.

routledge2

Routledge second measure.

routledge3

Routledge third measure. This is the exponential form of the second measure.

wilson

Wilson measure.

References

Cody, M. L. (1975). Towards a theory of continental species diversity: Bird distributions over Mediterranean habitat gradients. In M. L. Cody & J. M. Diamond (Eds.), Ecology and Evolution of Communities. Cambridge, MA: Harvard University Press, p. 214-257.

Routledge, R. D. (1977). On Whittaker's Components of Diversity. Ecology, 58(5), 1120-1127. DOI: 10.2307/1936932.

Whittaker, R. H. (1960). Vegetation of the Siskiyou Mountains, Oregon and California. Ecological Monographs, 30(3), 279-338. DOI: 10.2307/1943563.

Wilson, M. V., & Shmida, A. (1984). Measuring Beta Diversity with Presence-Absence Data. The Journal of Ecology, 72(3), 1055-1064. DOI: 10.2307/2259551.

See also

Examples

# Data from Magurran 1988, p. 162
trees <- IncidenceMatrix(
  data = c(1, 1, 1, 0, 0, 0,
           1, 1, 1, 1, 1, 1,
           0, 0, 1, 0, 1, 0,
           0, 0, 0, 1, 1, 1,
           0, 0, 0, 0, 1, 1,
           0, 0, 0, 1, 0, 1),
  nrow = 6, byrow = FALSE,
  dimnames = list(c("1", "2", "3", "4", "5", "6"),
                  c("Birch", "Oak", "Rowan", "Beech", "Hazel", "Holly"))
)

# Whittaker's measure
turnover(trees, "whittaker") # 1

#> $whittaker
#> [1] 1
#> 

# Cody's measure
turnover(trees, "cody") # 3

#> $cody
#> [1] 3
#> 

# Routledge's measures
turnover(trees, method = c("routledge1", "routledge2", "routledge3"),
         simplify = TRUE)

#> routledge1 routledge2 routledge3 
#>  0.2857143  0.5594978  1.7497936 

# 0.29 0.56 1.75

# Wilson and Shmida's measure
turnover(trees, "wilson") # 1

#> $wilson
#> [1] 1
#>