Zuni ceramics

A dataset containing ceramic counts from the Zuni region of the American Southwest.

zuni

Format

A numeric matrix with 420 rows (assemblages) and 18 variables (ceramic types). The numbers in brackets correspond to the date range of each type (in AD years):

LINO

Lino Gray (575-875).

KIAT

Kiatuthlanna Black-on-white (850-910).

RED

Red Mesa Black-on-white (900-1030).

GALL

Gallup Black-on-white (1025-1150).

ESC

Escavada Black-on-white (1050-1150).

PUBW

Puerco Black-on-white (1050-1200).

RES

Reserve Black-on-white (1071-1115).

TULA

Tularosa Black-on-white (1175-1300).

PINE

Pinedale Black-on-white (1275-1325).

PUBR

Puerco Black-on-red (1050-1200).

WING

Wingate Black-on-red (1070-1200).

WIPO

Wingate Polychrome (1150-1250).

SJ

St. Johns Black-on-red/Polychrome (1200-1300).

LSJ

St.Johns glaze, Techado Polychrome (1275-1300).

SPR

Springerville Polychrome (1250-1300).

PINER

Pinedale Black-on-red/Polychrome (1275-1325).

HESH

Heshotauthla Polychrome (1285-1400).

KWAK

Kwakina Polychrome (1285-1400).

Source

Peeples, M. A., & Schachner, G. (2012). Refining correspondence analysis-based ceramic seriation of regional data sets. Journal of Archaeological Science, 39(8), 2818-2827. DOI: 10.1016/j.jas.2012.04.040.