Heterogeneity and Evenness

index_heterogeneity() returns an heterogeneity or dominance index.

index_evenness() returns an evenness measure.

Usage

index_heterogeneity(object, ...)

index_evenness(object, ...)

# S4 method for CountMatrix
index_heterogeneity(
  object,
  method = c("berger", "brillouin", "mcintosh", "shannon", "simpson")
)

# S4 method for CountMatrix
index_evenness(
  object,
  method = c("shannon", "brillouin", "mcintosh", "simpson")
)

Arguments

object

A \(m \times p\) matrix of count data (typically a CountMatrix object).

...

Currently not used.

method

A character string specifying the index to be computed (see details). Any unambiguous substring can be given.

Value

Returns a DiversityIndex object.

Details

Diversity measurement assumes that all individuals in a specific taxa are equivalent and that all types are equally different from each other (Peet 1974). A measure of diversity can be achieved by using indices built on the relative abundance of taxa. These indices (sometimes referred to as non-parametric indices) benefit from not making assumptions about the underlying distribution of taxa abundance: they only take relative abundances of the species that are present and species richness into account. Peet (1974) refers to them as indices of heterogeneity.

Diversity indices focus on one aspect of the taxa abundance and emphasize either richness (weighting towards uncommon taxa) or dominance (weighting towards abundant taxa; Magurran 1988).

Evenness is a measure of how evenly individuals are distributed across the sample.

The following heterogeneity index and corresponding evenness measures are available (see Magurran 1988 for details):

berger

Berger-Parker dominance index. The Berger-Parker index expresses the proportional importance of the most abundant type. This metric is highly biased by sample size and richness, moreover it does not make use of all the information available from sample.

brillouin

Brillouin diversity index. The Brillouin index describes a known collection: it does not assume random sampling in an infinite population. Pielou (1975) and Laxton (1978) argues for the use of the Brillouin index in all circumstances, especially in preference to the Shannon index.

mcintosh

McIntosh dominance index. The McIntosh index expresses the heterogeneity of a sample in geometric terms. It describes the sample as a point of a \(S\)-dimensional hypervolume and uses the Euclidean distance of this point from the origin.

shannon

Shannon-Wiener diversity index. The Shannon index assumes that individuals are randomly sampled from an infinite population and that all taxa are represented in the sample (it does not reflect the sample size). The main source of error arises from the failure to include all taxa in the sample: this error increases as the proportion of species discovered in the sample declines (Peet 1974, Magurran 1988). The maximum likelihood estimator (MLE) is used for the relative abundance, this is known to be negatively biased by sample size.

simpson

Simpson dominance index for finite sample. The Simpson index expresses the probability that two individuals randomly picked from a finite sample belong to two different types. It can be interpreted as the weighted mean of the proportional abundances. This metric is a true probability value, it ranges from \(0\) (perfectly uneven) to \(1\) (perfectly even).

The berger, mcintosh and simpson methods return a dominance index, not the reciprocal or inverse form usually adopted, so that an increase in the value of the index accompanies a decrease in diversity.

Note

Ramanujan approximation is used for \(x!\) computation if \(x > 170\).

References

Berger, W. H. & Parker, F. L. (1970). Diversity of Planktonic Foraminifera in Deep-Sea Sediments. Science, 168(3937), 1345-1347. doi:10.1126/science.168.3937.1345 .

Brillouin, L. (1956). Science and information theory. New York: Academic Press.

Kintigh, K. W. (1989). Sample Size, Significance, and Measures of Diversity. In Leonard, R. D. and Jones, G. T., Quantifying Diversity in Archaeology. New Directions in Archaeology. Cambridge: Cambridge University Press, p. 25-36.

Laxton, R. R. (1978). The measure of diversity. Journal of Theoretical Biology, 70(1), 51-67. doi:10.1016/0022-5193(78)90302-8 .

Magurran, A. E. (1988). Ecological Diversity and its Measurement. Princeton, NJ: Princeton University Press. doi:10.1007/978-94-015-7358-0 .

McIntosh, R. P. (1967). An Index of Diversity and the Relation of Certain Concepts to Diversity. Ecology, 48(3), 392-404. doi:10.2307/1932674 .

Peet, R. K. (1974). The Measurement of Species Diversity. Annual Review of Ecology and Systematics, 5(1), 285-307. doi:10.1146/annurev.es.05.110174.001441 .

Pielou, E. C. (1975). Ecological Diversity. New York: Wiley. doi:10.4319/lo.1977.22.1.0174b

Shannon, C. E. (1948). A Mathematical Theory of Communication. The Bell System Technical Journal, 27, 379-423. doi:10.1002/j.1538-7305.1948.tb01338.x .

Simpson, E. H. (1949). Measurement of Diversity. Nature, 163(4148), 688-688. doi:10.1038/163688a0 .

See also

Author

N. Frerebeau

Examples

## Coerce dataset to a count matrix
data("chevelon", package = "folio")
chevelon <- as_count(chevelon)

## Shannon diversity index
(index_h <- index_heterogeneity(chevelon, method = "shannon"))
#> <HeterogeneityIndex: shannon>
#>       size     index
#> P610s   14 1.6681740
#> P610e    8 1.2130076
#> P625     4 0.6931472
#> P630     1 0.0000000
#> P307     1 0.0000000
#> P631    12 1.7481555
#> P623     3 1.0986123
#> P624    14 1.8711604
#> P626s   26 1.7207095
#> P626e  106 1.9115521
#> P627     4 1.0397208
#> P628    12 1.8200760
(index_e <- index_evenness(chevelon, method = "shannon"))
#> <EvennessIndex: shannon>
#>       size     index
#> P610s   14 0.8572718
#> P610e    8 0.8750000
#> P625     4 1.0000000
#> P630     1       NaN
#> P307     1       NaN
#> P631    12 0.8983742
#> P623     3 1.0000000
#> P624    14 0.9615862
#> P626s   26 0.8842698
#> P626e  106 0.8699849
#> P627     4 0.9463946
#> P628    12 0.9353340

## Bootstrap resampling
bootstrap(index_e)
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>             min      mean       max     lower     upper       Q25       Q50
#> P610s 0.5970948 0.8728523 0.9926141 0.8692697 0.8764349 0.8312828 0.8771407
#> P610e 0.5435644 0.8684110 1.0000000        NA        NA 0.8194484 0.8868595
#> P625  0.8112781 0.8956327 1.0000000        NA        NA 0.8112781 0.8112781
#> P630        Inf       NaN      -Inf        NA        NA        NA        NA
#> P307        Inf       NaN      -Inf        NA        NA        NA        NA
#> P631  0.5152734 0.8925811 1.0000000 0.8890283 0.8961339 0.8592975 0.8979445
#> P623  0.9182958 0.9392715 1.0000000        NA        NA 0.9182958 0.9182958
#> P624  0.7544450 0.9208799 0.9931223 0.9184137 0.9233462 0.8982265 0.9299064
#> P626s 0.7240480 0.8861452 0.9934151 0.8834071 0.8888833 0.8599017 0.8898776
#> P626e 0.7685511 0.8718347 0.9487089 0.8699662 0.8737032 0.8515753 0.8728703
#> P627  0.8112781 0.9096148 1.0000000        NA        NA 0.8112781 0.9463946
#> P628  0.6826062 0.9136603 1.0000000 0.9108150 0.9165057 0.8848617 0.9155385
#>             Q75
#> P610s 0.9134443
#> P610e 0.9463946
#> P625  1.0000000
#> P630         NA
#> P307         NA
#> P631  0.9355245
#> P623  1.0000000
#> P624  0.9546444
#> P626s 0.9169214
#> P626e 0.8936265
#> P627  0.9463946
#> P628  0.9439593

## Jackknife resampling
jackknife(index_e)
#>            mean          bias        error
#> P610s 0.8530134 -3.832613e-02 3.467900e-02
#> P610e 0.8725916 -2.167581e-02 1.055074e-01
#> P625        NaN           NaN          NaN
#> P630        NaN           NaN          NaN
#> P307        NaN           NaN          NaN
#> P631  0.8956783 -2.426312e-02 3.790426e-02
#> P623  1.0000000  9.992007e-16 3.330669e-16
#> P624  0.9598834 -1.532543e-02 1.972936e-02
#> P626s 0.8786773 -5.033240e-02 4.931785e-02
#> P626e 0.8645734 -4.870273e-02 4.770957e-02
#> P627  0.9461354 -2.333001e-03 6.329892e-02
#> P628  0.9329141 -2.177978e-02 1.598781e-02