Jackknife Estimation

Jackknife Estimation

Usage

# S4 method for DiversityIndex
jackknife(object, f = NULL)

Arguments

object

An R object (typically a DiversityIndex object).

f

A function that takes a single numeric vector (the leave-one-out values of do) as argument.

Value

If f is NULL (the default), jackknife() returns a named numeric

vector with the following elements:

original

The observed value of do applied to object.

mean

The jackknife estimate of mean of do.

bias

The jackknife estimate of bias of do.

error

he jackknife estimate of standard error of do.

If f is a function, jackknife() returns the result of f applied to the leave-one-out values of do.

See also

Author

N. Frerebeau

Examples

## Data from Conkey 1980, Kintigh 1989
data("cantabria")

## Shannon diversity index
(h <- heterogeneity(cantabria, method = "shannon"))
#> [1] 3.269200 2.955298 2.491683 2.485604 2.329187

## Bootstrap resampling
bootstrap(h, f = NULL)
#>                  original     mean         bias     error
#> Altamira         3.269200 3.279013  0.009813698 0.0817979
#> Cueto de la Mina 2.955298 2.973579  0.018280757 0.1480563
#> El Juyo          2.491683 2.531218  0.039535424 0.2003939
#> El Cierro        2.485604 2.496113  0.010508789 0.2202203
#> La Paloma        2.329187 2.308531 -0.020655914 0.2672030

bootstrap(h, f = summary)
#>                      Min.  1st Qu.   Median     Mean  3rd Qu.     Max.
#> Altamira         2.899258 3.226897 3.281514 3.276767 3.332574 3.497370
#> Cueto de la Mina 2.454262 2.872974 2.975768 2.972146 3.079099 3.342932
#> El Juyo          1.810996 2.390225 2.528202 2.522168 2.661989 3.050147
#> El Cierro        1.681235 2.360348 2.522090 2.503370 2.658155 3.059080
#> La Paloma        1.213008 2.173499 2.344600 2.319279 2.480434 3.000357

quant <- function(x) quantile(x, probs = c(0.25, 0.50))
bootstrap(h, f = quant)
#>                       25%      50%
#> Altamira         3.225558 3.281150
#> Cueto de la Mina 2.883125 2.981846
#> El Juyo          2.395470 2.520553
#> El Cierro        2.332171 2.500438
#> La Paloma        2.149637 2.331647

## Jackknife resampling
jackknife(h)
#>                  original     mean       bias      error
#> Altamira         3.269200 3.246457 -0.9779158 0.08145529
#> Cueto de la Mina 2.955298 2.932856 -0.9649994 0.15907033
#> El Juyo          2.491683 2.469588 -0.9500688 0.19670002
#> El Cierro        2.485604 2.462723 -0.9838709 0.22195573
#> La Paloma        2.329187 2.305876 -1.0023704 0.25266291

bootstrap(h, f = summary)
#>                      Min.  1st Qu.   Median     Mean  3rd Qu.     Max.
#> Altamira         2.949940 3.230484 3.285142 3.281061 3.339651 3.491809
#> Cueto de la Mina 2.367749 2.873841 2.978449 2.973666 3.086736 3.376053
#> El Juyo          1.555410 2.414556 2.534501 2.526948 2.655935 3.070988
#> El Cierro        1.424130 2.350502 2.503393 2.489803 2.652815 3.101226
#> La Paloma        1.153742 2.141848 2.345561 2.310420 2.505769 2.940143