Model Refining

Model Refining

refine(object, ...)

# S4 method for CountMatrix
refine(object, cutoff, n = 1000, axes = c(1,
  2), ...)

# S4 method for DateModel
refine(object, method = c("jackknife",
  "bootstrap"), n = 1000, ...)

Arguments

object
An \(m \times p\) data matrix.

...
Further arguments to be passed to internal methods.

cutoff
A function that takes a numeric vector as argument and returns a single numeric value (see details).

n
A non-negative integer giving the number of partial bootstrap replications (see details).

axes
A numeric vector giving the subscripts of the CA axes to use (see details).

method
A character string specifiying the resampling method to be used. This must be one of "jackknife", "bootstrap" (see details). Any unambiguous substring can be given.

Value

TODO

Methods (by class)

CountMatrix: Performs a partial bootstrap correspondance analysis seriation refinement and returns a BootCA object.

DateModel: Checks the stability of a date model with resampling methods.

Note

These methods can lead to much longer execution times and larger output objects. To monitor the execution of these re-sampling procedures, a progress bar will automatically be displayed if the pbapply package is installed on your machine.

CA seriation refining

refine allows to identify samples that are subject to sampling error or samples that have underlying structural relationships and might be influencing the ordering along the CA space.

This relies on a partial bootstrap approach to CA-based seriation where each sample is replicated n times. The maximum dimension length of the convex hull around the sample point cloud allows to remove samples for a given cutoff value.

According to Peebles and Schachner (2012), "[this] point removal procedure [results in] a reduced dataset where the position of individuals within the CA are highly stable and which produces an ordering consistend with the assumptions of frequency seriation."

If the results of refine is used as an input argument in seriate, a correspondance analysis is performed on the subset of object which matches the samples to be kept. Then excluded samples are projected onto the dimensions of the CA coordinate space using the row transition formulae. Finally, row coordinates onto the first dimension give the seriation order.

Date model checking

Resampling methods can be used to check the stability of the resulting DateModel.

If jackknife is used, one type/fabric is removed at a time and all statistics are recalculated. In this way, one can assess whether certain type/fabric has a substantial influence on the date estimate. A six columns data.frame is returned, giving the results of the resamping procedure (jackknifing fabrics) for each assemblage (in rows) with the following columns:

id

An identifier to link each row to an assemblage.

date

The jackknife event date estimate.

lower

The lower boundary of the associated prediction interval.

upper

The upper boundary of the associated prediction interval.

error

The standard error of predicted means.

bias

The jackknife estimate of bias.

If bootstrap is used, a large number of new bootstrap assemblages is created, with the same sample size, by resampling each of the original assemblage with replacement. Then, examination of the bootstrap statistics makes it possible to pinpoint assemblages that require further investigation. A six columns data.frame is returned, giving the boostrap distribution statistics for each replicated assemblage (in rows) with the following columns:

id

An identifier to link each row to an assemblage.

min

Minimum value.

Q05

Sample quantile to 0.05 probability.

mean

Mean value (event date).

Q95

Sample quantile to 0.95 probability.

max

Maximum value.

References

Bellanger, L., Tomassone, R. & Husi, P. (2008). A Statistical Approach for Dating Archaeological Contexts. Journal of Data Science, 6, 135-154.

Bellanger, L., Husi, P. & Tomassone, R. (2006). Une approche statistique pour la datation de contextes archéologiques. Revue de Statistique Appliquée, 54(2), 65-81.

Peeples, M. A., & Schachner, G. (2012). Refining correspondence analysis-based ceramic seriation of regional data sets. Journal of Archaeological Science, 39(8), 2818-2827. DOI: 10.1016/j.jas.2012.04.040.

See also