Resampling Methods

resample() simulate observations from a multinomial distribution.

bootstrap() generate bootstrap estimations of a statistic.

jackknife() generate jackknife estimations of a statistic.

Usage

resample(object, ...)

# S4 method for DiversityIndex
bootstrap(object, n = 1000, f = NULL)

# S4 method for DiversityIndex
jackknife(object)

# S4 method for numeric
resample(object, do, n, size = sum(object), ..., f = NULL)

Arguments

object

A numeric vector of count data (absolute frequencies).

...

Extra arguments passed to do.

n

A non-negative integer specifying the number of bootstrap replications.

f

A function that takes a single numeric vector (the result of do) as argument.

do

A function that takes object as an argument and returns a single numeric value.

size

A non-negative integer specifying the sample size.

Value

If f is NULL, resample() returns the n values of do. Else, returns the result of f applied to the n values of do.

If f is NULL, bootstrap() and jackknife() return a data.frame

with the following elements (else, returns the result of f applied to the n values of do) :

original

The observed value of do applied to object.

mean

The bootstrap/jackknife estimate of mean of do.

bias

The bootstrap/jackknife estimate of bias of do.

error

The boostrap/jackknife estimate of standard error of do.

See also

Author

N. Frerebeau

Examples

## Sample observations from a multinomial distribution
x <- sample(1:100, 50, TRUE)
resample(x, do = median, n = 100)
#>   [1] 55.5 58.0 52.0 54.0 54.0 57.5 52.0 52.5 52.0 53.0 58.5 56.0 53.0 53.0 49.5
#>  [16] 59.0 54.0 56.5 52.5 50.5 54.5 53.5 55.5 55.0 55.5 52.5 50.0 57.0 56.0 54.0
#>  [31] 52.0 53.5 58.5 50.5 54.0 55.0 54.0 50.5 55.0 52.5 54.5 51.0 54.0 53.0 55.5
#>  [46] 52.0 56.0 58.5 56.0 56.5 53.0 55.0 52.0 56.0 55.5 59.0 57.5 55.5 55.0 57.0
#>  [61] 59.0 54.0 56.5 51.5 58.0 53.5 59.5 50.5 54.5 58.0 53.0 54.5 56.0 60.0 52.0
#>  [76] 54.0 55.5 54.0 57.5 56.5 55.5 56.5 55.0 61.0 54.5 53.5 61.0 54.0 53.0 55.0
#>  [91] 53.5 52.0 53.0 57.5 54.5 53.5 53.0 54.5 55.0 58.5

## Estimate the 25th, 50th and 95th percentiles
quant <- function(x) { quantile(x, probs = c(0.25, 0.50, 0.75)) }
resample(x, n = 100, do = median, f = quant)
#> 25% 50% 75% 
#>  53  54  56