tabula: Analysis, Seriation and Visualization of Archaeological Count Data

An easy way to examine archaeological count data (artifacts, faunal remains, etc.). This package includes several measures of diversity: e.g. richness and rarefaction (Chao1, Chao2, ACE, ICE, etc.), diversity/dominance and evenness (Brillouin, Shannon, Simpson, etc.), turnover and similarity (Brainerd-Robinson, ...). Most of these methods are described and discussed in Maguran (1988) <doi:10.1007/978-94-015-7358-0>. It also provides matrix seriation methods (reciprocal ranking, CA-based seriation) for chronological modeling and dating. The package make it easy to visualize count data and statistical thresholds: rank/abundance plots, Ford (1972) <isbn:0913134082> and Bertin (1977) <isbn:2082111121> diagrams, etc.

Details

Package:

tabula

Type:

Package

Version:

1.2.0

Date:

2019-03-19

License:

GPL-3

DOI:

10.5281/zenodo.1489944

Note

This work received a state financial support managed by the Agence Nationale de la Recherche (France) throught the program Investissements d'avenir (ref. ANR-10-LABX-52).

See also